feat: drop jakarta validation (#76)
* fix: drop jakarta validation. Issue #73

* test: fix test cases

* docs: update changelog

* chore: fix login issue and update reachability metadata

* chore: redact all password records

* feat: make all validations the same messages as jakarta

* test: add test cases for missing scenarios

* chore: refactor validation

* fix: date validation throwing errors

* chore: updated reachability json file and added scripts for testing

* chore: make usage of same patter for validation

+++ server/src/main/java/br/com/tasknoteapp/server/controller/UserSessionController.java
@@ -1,13 +1,16 @@
 package br.com.tasknoteapp.server.controller;
 
 import br.com.tasknoteapp.server.entity.User;
+import br.com.tasknoteapp.server.exception.InternalValidationException;
+import br.com.tasknoteapp.server.exception.InvalidCredentialsException;
 import br.com.tasknoteapp.server.exception.UserNotFoundException;
 import br.com.tasknoteapp.server.request.DeleteAccountRequest;
 import br.com.tasknoteapp.server.response.JwtAuthenticationResponse;
 import br.com.tasknoteapp.server.response.UserResponse;
 import br.com.tasknoteapp.server.service.AuthService;
 import br.com.tasknoteapp.server.service.UserSessionService;
-import jakarta.validation.Valid;
+import br.com.tasknoteapp.server.util.ValidationUtil;
+import java.util.Optional;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -47,11 +50,25 @@ public class UserSessionController {
    * @return {@link UserResponse} with the user information.
    */
   @PostMapping("/delete-account")
-  public ResponseEntity<UserResponse> deleteAccount(
-      @RequestBody @Valid DeleteAccountRequest request) {
+  public ResponseEntity<UserResponse> deleteAccount(@RequestBody DeleteAccountRequest request) {
+    Optional<InternalValidationException> deleteValidation = isDeleteAccountRequestValid(request);
+    if (deleteValidation.isPresent()) {
+      throw new InvalidCredentialsException();
+    }
     User user = authService.getCurrentUser().orElseThrow(UserNotFoundException::new);
     authService.verifyCurrentPassword(user, request.password());
     UserResponse deleted = userSessionService.deleteCurrentUserAccount();
     return ResponseEntity.ok(deleted);
+  }
+
+  private Optional<InternalValidationException> isDeleteAccountRequestValid(
+      DeleteAccountRequest request) {
+    try {
+      ValidationUtil.notNullNorBlank("password", request.password());
+      ValidationUtil.maxSize("password", request.password(), ValidationUtil.MAX_PASSWORD_SIZE);
+      return Optional.empty();
+    } catch (InternalValidationException ex) {
+      return Optional.of(ex);
+    }
   }
 }
